이미지연습/requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.4
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
firebase-admin==4.5.1
Flask==1.1.2
future==0.18.2
gcloud==0.17.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jwcrypto==0.8
jws==0.1.3
keyboard==0.13.5
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
numpy==1.20.1
oauth2client==3.0.0
openpyxl==3.0.6
packaging==20.9
pandas==1.2.2
pefile==2019.4.18
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pygame==2.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.6.0
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7
schedule==1.0.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sseclient==0.0.27
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
